refactor(core): group feats, proficiencies, and progression packages
Move fragmented modules into core/feats/, core/proficiencies/, and
core/progression/ with facades preserved via package __init__ exports.
Keep core/levels.py at package root to avoid circular imports with models.
