+++ ui/menus/feats/_level_up.py
@@ -2,13 +2,15 @@
 
 from typing import Any
 
-from core.asi import cap_stats
-from core.feat_visibility import build_feat_selection_context_from_character
 from core.feats import (
     list_feats_for_selection,
     resolve_feat_ability_bonuses,
 )
+from core.feats.feat_visibility import (
+    build_feat_selection_context_from_character,
+)
 from core.localization import get_string
+from core.progression.asi import cap_stats
 from core.types import StatMap, StringsDict
 from ui.menus import _deps
 from ui.menus._common import _print_screen_header
@@ -29,7 +31,7 @@ def select_level_up_feat_or_asi(
     """
     from dataclasses import replace
 
-    from core.asi import apply_asi_one_two, apply_asi_two_one
+    from core.progression.asi import apply_asi_one_two, apply_asi_two_one
     from ui.menus.asi import select_asi_mode, select_asi_stats
 
     _print_screen_header(
